+++ test/stochastic_characterisation.jl
@@ -25,7 +25,7 @@ f!(du, u, p, t) = du .= 1.01u
     @test CoupledSDEs(sciml_prob(corr)).noise_type == types
 
     # invertible correlated Wiener noise
-    g!(du, u, p, t) = (du .= [1 0.3; 0.3 1]; return nothing)
+    g! = (du, u, p, t) -> (du .= [1 0.3; 0.3 1]; return nothing)
     corr_alt = CoupledSDEs(f!, zeros(2); g = g!, noise_prototype = zeros(2, 2))
     types = corr_alt.noise_type
     @test values(types) == (true, true, true, true)
@@ -38,14 +38,14 @@ f!(du, u, p, t) = du .= 1.01u
     @test CoupledSDEs(sciml_prob(corr_noninv)).noise_type == types
 
     # non-invertible correlated Wiener noise
-    g!(du, u, p, t) = (du .= [1 0.3 1; 0.3 1 1]; return nothing)
+    g! = (du, u, p, t) -> (du .= [1 0.3 1; 0.3 1 1]; return nothing)
     corr_alt = CoupledSDEs(f!, zeros(2); g = g!, noise_prototype = zeros(2, 3))
     types = corr_alt.noise_type
     @test values(types) == (true, true, true, false)
     @test CoupledSDEs(sciml_prob(corr_alt)).noise_type == types
 
     # non-autonomous Wiener noise
-    g!(du, u, p, t) = (du .= 1 / (1 + t); return nothing)
+    g! = (du, u, p, t) -> (du .= 1 / (1 + t); return nothing)
     addit_non_autom = CoupledSDEs(f!, zeros(2); g = g!)
     types = addit_non_autom.noise_type
     @test values(types) == (true, false, true, false)
@@ -54,7 +54,7 @@ end
 
 @testset "Multiplicative noise" begin
     # multiplicative linear Wiener noise
-    g!(du, u, p, t) = (du .= u; return nothing)
+    g! = (du, u, p, t) -> (du .= u; return nothing)
     linear_multipli = CoupledSDEs(f!, rand(2) ./ 10; g = g!)
     types = linear_multipli.noise_type
     @test values(types) == (false, true, true, false)
@@ -67,21 +67,21 @@ end
     @test CoupledSDEs(sciml_prob(lin_multipli_alt)).noise_type == types
 
     # multiplicative nonlinear Wiener noise
-    g!(du, u, p, t) = (du .= u .^ 2; return nothing)
+    g! = (du, u, p, t) -> (du .= u .^ 2; return nothing)
     nonlinear_multiplicative = CoupledSDEs(f!, rand(2); g = g!)
     types = nonlinear_multiplicative.noise_type
     @test values(types) == (false, true, false, false)
     @test CoupledSDEs(sciml_prob(nonlinear_multiplicative)).noise_type == types
 
     # non-autonomous linear multiplicative Wiener noise
-    g!(du, u, p, t) = (du .= u ./ (1 + t); return nothing)
+    g! = (du, u, p, t) -> (du .= u ./ (1 + t); return nothing)
     addit_non_autom = CoupledSDEs(f!, zeros(2); g = g!)
     types = addit_non_autom.noise_type
     @test values(types) == (false, false, true, false)
     @test CoupledSDEs(sciml_prob(addit_non_autom)).noise_type == types
 
     # non-autonomous nonlinear multiplicative Wiener noise
-    g!(du, u, p, t) = (du .= u .^ 2 ./ (1 + t); return nothing)
+    g! = (du, u, p, t) -> (du .= u .^ 2 ./ (1 + t); return nothing)
     addit_non_autom = CoupledSDEs(f!, zeros(2); g = g!)
     types = addit_non_autom.noise_type
     @test values(types) == (false, false, false, false)